services/pdf-parser/requirements.txt
flask==3.0.0
pdfplumber==0.11.0
camelot-py[cv]==0.11.0
pandas==2.1.4
opencv-python==4.8.1.78
Werkzeug==3.0.1
gunicorn==21.2.0
